updated communication code and test code

+++ DrivetrainFirmware/src/motor_control.cpp
@@ -23,7 +23,7 @@ AccelStepper* get_motor(MotorID id){
 }
 
 // drive motor by position amount
-void drive_motor(MotorID id, long position){
+void drive_motor(MotorID id, int position){
     AccelStepper* motor = get_motor(id);
     if(motor != nullptr){
         motor->setCurrentPosition(0);
@@ -35,7 +35,7 @@ void drive_motor(MotorID id, long position){
 }
 
 // drive both motors forward by position amount
-void drive_forward(long position){
+void drive_forward(int position){
     Left_Motor.setCurrentPosition(0);
     Right_Motor.setCurrentPosition(0);
     Left_Motor.moveTo(position);
@@ -47,7 +47,7 @@ void drive_forward(long position){
 }   
 
 // drive motors independently
-void mix_drive(long left_position, long right_position){
+void mix_drive(int left_position, int right_position){
     Left_Motor.setCurrentPosition(0);
     Right_Motor.setCurrentPosition(0);
     Left_Motor.moveTo(left_position);
@@ -59,19 +59,19 @@ void mix_drive(long left_position, long right_position){
 }
 
 // convert mm to steps
-long mm_to_steps(long mm){
-    return static_cast<long>(mm * STEPS_PER_MM);
+int mm_to_steps(int mm){
+    return static_cast<int>(mm * STEPS_PER_MM);
 }
 
 // calculate steps for a turn in place
 // direction: + -> turn right, - -> turn left
-void calculate_turn_steps(long turn_angle_deg,
-    long* left_steps, long* right_steps){
+void calculate_turn_steps(int turn_angle_deg,
+    int* left_steps, int* right_steps){
     // calculate the arc length each wheel needs to travel
     double turn_circumference_mm = WHEEL_CENTER_DISTANCE_MM * 3.14159; // assuming 200mm between wheels
     double arc_length_mm = (turn_angle_deg / 360.0) * turn_circumference_mm;    
     // convert to steps
-    long steps = mm_to_steps(static_cast<long>(arc_length_mm));
+    int steps = mm_to_steps(static_cast<int>(arc_length_mm));
     // set left and right steps (one forward, one backward)
     *left_steps = steps;
     *right_steps = -steps;
@@ -79,15 +79,15 @@ void calculate_turn_steps(long turn_angle_deg,
 
 // turn robot in place by angle
 // direction: + -> turn right, - -> turn left
-void turn_in_place(long turn_angle_deg){
-    long left_steps = 0;
-    long right_steps = 0;
+void turn_in_place(int turn_angle_deg){
+    int left_steps = 0;
+    int right_steps = 0;
     calculate_turn_steps(turn_angle_deg, &left_steps, &right_steps);
     mix_drive(left_steps, right_steps);
 }
 
 // drive both motors forward by distance in mm
-void drive_forward_distance_mm(long distance_mm){
-    long steps = mm_to_steps(distance_mm);
+void drive_forward_distance_mm(int distance_mm){
+    int steps = mm_to_steps(distance_mm);
     drive_forward(steps);
 }
